Cria função para fechar o modal ao teclar ESC
